+++ tests/libtest/lib1512.c
@@ -40,7 +40,7 @@ static CURLcode test_lib1512(char *URL)
   char *address = libtest_arg2;
   char dnsentry[256];
   struct curl_slist *slist = NULL;
-  int i;
+  size_t i;
   char target_url[256];
   (void)URL; /* URL is setup in the code */
 
@@ -61,7 +61,7 @@ static CURLcode test_lib1512(char *URL)
     /* specify target */
     curl_msnprintf(target_url, sizeof(target_url),
                    "http://server.example.curl:%s/path/1512%04i",
-                   port, i + 1);
+                   port, (int)i + 1);
     target_url[sizeof(target_url) - 1] = '\0';
     easy_setopt(curl[i], CURLOPT_URL, target_url);
     /* go verbose */
